% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataRules.r
\name{apply_export_data_rules}
\alias{apply_export_data_rules}
\title{Apply final export rules to EHR-BSI tables}
\usage{
apply_export_data_rules(result_list)
}
\arguments{
\item{result_list}{List containing ehrbsi, patient, isolate, res tables}
}
\value{
Updated result list after export rules are applied
}
\description{
Collates last-step data checks and standardizations before output is returned.
}
