% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{apply_lookup_chain}
\alias{apply_lookup_chain}
\title{Apply a chain of lookups to a column}
\usage{
apply_lookup_chain(
  data,
  column_name,
  lookup_chain,
  track_unmapped = TRUE,
  fallback_value = NULL
)
}
\arguments{
\item{data}{Data frame containing the column to recode}

\item{column_name}{Name of column to recode}

\item{lookup_chain}{List of lookup vectors to apply in sequence}

\item{track_unmapped}{Whether to track and report unmapped values}

\item{fallback_value}{Value or function to use for unmapped items}
}
\value{
Data frame with recoded column
}
\description{
Apply a chain of lookups to a column
}
