% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_to_all_tables}
\alias{apply_to_all_tables}
\title{Apply a transformation function to all standard EHR-BSI tables}
\usage{
apply_to_all_tables(result_list, transform_fn)
}
\arguments{
\item{result_list}{List containing ehrbsi, patient, isolate, res tables}

\item{transform_fn}{Function to apply to each table}
}
\value{
Modified result list
}
\description{
Apply a transformation function to all standard EHR-BSI tables
}
