% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_pathogen_colors}
\alias{assign_pathogen_colors}
\title{Assign colors to pathogens, generating extras for unrecognized names}
\usage{
assign_pathogen_colors(pathogen_names)
}
\arguments{
\item{pathogen_names}{Character vector of pathogen names}
}
\value{
Named character vector of hex colors
}
\description{
Assign colors to pathogens, generating extras for unrecognized names
}
