% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{auto_detect_lookups}
\alias{auto_detect_lookups}
\title{Auto-detect lookups from Lookups tab}
\usage{
auto_detect_lookups(country_code)
}
\description{
Auto-detect lookups from Lookups tab
}
\keyword{internal}
