% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{build_ab_table}
\alias{build_ab_table}
\title{Build antibiogram summary table}
\usage{
build_ab_table(rctx_df)
}
\arguments{
\item{rctx_df}{Resistance data with context}
}
\value{
Data frame with antibiogram summary
}
\description{
Build antibiogram summary table
}
