% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{build_antibiotic_config}
\alias{build_antibiotic_config}
\title{Build antibiotic configuration from Excel config}
\usage{
build_antibiotic_config(excel_config)
}
\description{
Build antibiotic configuration from Excel config
}
\keyword{internal}
