% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{build_defaults_from_excel}
\alias{build_defaults_from_excel}
\title{Build table defaults from Excel config}
\usage{
build_defaults_from_excel(excel_config, table_name)
}
\description{
Build table defaults from Excel config
}
\keyword{internal}
