% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{build_lookup_mappings_from_excel}
\alias{build_lookup_mappings_from_excel}
\title{Build lookup mappings from Excel config}
\usage{
build_lookup_mappings_from_excel(excel_config)
}
\description{
Build lookup mappings from Excel config
}
\keyword{internal}
