% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{build_special_fields_from_excel}
\alias{build_special_fields_from_excel}
\title{Build special fields from Excel config}
\usage{
build_special_fields_from_excel(excel_config)
}
\description{
Build special fields from Excel config
}
\keyword{internal}
