% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_unit_id}
\alias{compute_unit_id}
\title{Compute UnitId from HospitalId and UnitSpecialtyShort}
\usage{
compute_unit_id(data)
}
\arguments{
\item{data}{Data frame with HospitalId and UnitSpecialtyShort columns}
}
\value{
Character vector of UnitId values
}
\description{
Shared logic used by EE and GEN country transforms.
}
