% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configWizard.R
\name{config_wizard_app}
\alias{config_wizard_app}
\title{EHR-BSI Country Config Wizard (Shiny)}
\usage{
config_wizard_app()
}
\description{
A guided assistant to create or migrate country configuration files
with minimal manual editing. Supports schema v2 export.
}
