% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\docType{data}
\name{COUNTRY_R_TRANSFORMS}
\alias{COUNTRY_R_TRANSFORMS}
\title{Country R-Only Transformations}
\format{
List of R-only transformations that cannot be defined in Excel
}
\usage{
COUNTRY_R_TRANSFORMS
}
\description{
Country R-Only Transformations
}
\note{
This contains field_transforms functions that require R code execution.
All other configuration should be in the Excel Config tab.
}
\keyword{datasets}
