% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{create_age_groups}
\alias{create_age_groups}
\title{Create age groups from numeric ages}
\usage{
create_age_groups(ages)
}
\arguments{
\item{ages}{Numeric vector of ages}
}
\value{
Factor vector with age group categories
}
\description{
Create age groups from numeric ages
}
