% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{create_flexible_record_ids}
\alias{create_flexible_record_ids}
\title{Create flexible record IDs based on templates}
\usage{
create_flexible_record_ids(data, id_templates, config)
}
\arguments{
\item{data}{Data frame}

\item{id_templates}{Named list of ID templates (bsi, patient, isolate)}

\item{config}{Country configuration object}
}
\value{
Data frame with record ID columns added
}
\description{
Create flexible record IDs based on templates
}
