% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{create_lookup_vector}
\alias{create_lookup_vector}
\title{Create lookup vector from lookup table}
\usage{
create_lookup_vector(lookup_table, value_col, key_col)
}
\arguments{
\item{lookup_table}{Data frame with lookup values}

\item{value_col}{Name of column containing target values}

\item{key_col}{Name of column containing source keys}
}
\value{
Named vector for use with dplyr::recode
}
\description{
Create lookup vector from lookup table
}
