% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{create_standard_patient_table}
\alias{create_standard_patient_table}
\title{Create standard patient table structure}
\usage{
create_standard_patient_table(
  data,
  record_id_col = "record_id_patient",
  parent_id_col = "record_id_bsi",
  country_defaults = list()
)
}
\arguments{
\item{data}{Source data frame}

\item{record_id_col}{Name of patient record ID column}

\item{parent_id_col}{Name of parent (BSI) record ID column}

\item{country_defaults}{List of country-specific default values}
}
\value{
Data frame with standard patient table structure
}
\description{
Create standard patient table structure
}
