% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_time_period}
\alias{create_time_period}
\title{Create time period from dates based on aggregation level}
\usage{
create_time_period(dates, aggregation = "month")
}
\arguments{
\item{dates}{Date vector}

\item{aggregation}{Character: "week", "month", "quarter", or "year"}
}
\value{
Date vector with period start dates
}
\description{
Create time period from dates based on aggregation level
}
