% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingEstonia.R
\name{.process_estonia_basic_cleaning}
\alias{.process_estonia_basic_cleaning}
\title{Process Estonia BSI data from raw format to EHR-BSI format}
\usage{
.process_estonia_basic_cleaning(raw_data, reporting_year)
}
\arguments{
\item{raw_data}{Received from genericRecodeOrchestrator.R}
}
\value{
Returns a list containing the four EHR-BSI data tables: ehrbsi, patient, isolate, res
}
\description{
Process Estonia BSI data from raw format to EHR-BSI format
}
\note{
The reporting year is automatically derived from the admission dates in the data
}
