% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingMalta.R
\name{.process_malta_basic_cleaning}
\alias{.process_malta_basic_cleaning}
\title{Process Malta BSI data from raw format to EHR-BSI format}
\usage{
.process_malta_basic_cleaning(raw_data, reporting_year)
}
\arguments{
\item{raw_data}{Received from genericRecodeOrchestrator.R}
}
\value{
Returns a list containing the four EHR-BSI data tables: ehrbsi, patient, isolate, res
}
\description{
Process Malta BSI data from raw format to EHR-BSI format
}
