% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHRBSI_Convertor.R
\name{EHRBSI_Convertor}
\alias{EHRBSI_Convertor}
\title{Function to convert data from EARS-Net format into EHR-BSI format}
\usage{
EHRBSI_Convertor(
  input_file = NULL,
  input_file_path = NULL,
  write_to_file = FALSE,
  write_to_file_path = NULL,
  reporting_unit = "Both"
)
}
\arguments{
\item{input_file_path}{Path to the input file, defaults to the working directory}

\item{write_to_file}{Whether to write the output list into separate output CSV-files}

\item{write_to_file_path}{File path where to write the files, defaults to the working directory}

\item{reporting_unit}{Reporting unit: "Hospital" is recommended, also accepts "Laboratory" or "Both" (Hospital and laboratory), or "Country"}
}
\value{
Returns a list of objects (data.tables) in EHR-BSI format
}
\description{
Function to convert data from EARS-Net format into EHR-BSI format
}
\examples{
EHRBSI_Convertor(input_file = "ADD_TEST.csv")
}
