% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{find_tolerant_column_match}
\alias{find_tolerant_column_match}
\title{Find tolerant column name match}
\usage{
find_tolerant_column_match(raw_names, data_names, norm_func)
}
\arguments{
\item{raw_names}{Character vector of raw column names to find}

\item{data_names}{Character vector of available column names in data}

\item{norm_func}{Function to normalize names for comparison}
}
\value{
Named list with matches (name = raw_name, value = matched_data_name)
}
\description{
Find tolerant column name match
}
