% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{format_date_for_id}
\alias{format_date_for_id}
\title{Format date column for use in record IDs}
\usage{
format_date_for_id(date_col, format_type = "date")
}
\arguments{
\item{date_col}{Date vector to format}

\item{format_type}{Type of formatting: "date", "datetime", or "year"}
}
\value{
Character vector of formatted dates
}
\description{
Format date column for use in record IDs
}
