% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/From_EHRBSI_Convertor.R
\name{From_EHRBSI_converter}
\alias{From_EHRBSI_converter}
\title{Function to convert data from EHR-BSI format to EARS-Net format}
\usage{
From_EHRBSI_converter(
  input_file = NULL,
  input_file_path = NULL,
  write_to_file = FALSE,
  write_to_file_path = NULL
)
}
\arguments{
\item{input_file}{Input CSV files in EHR-BSI format, please give a vector of names for all 4 files "EHRBSI", "Patient", "Isolate" and "Res", defaults to TESSy file names}

\item{input_file_path}{Path to the input files, defaults to the working directory}

\item{write_to_file}{Whether to write the output list into separate output CSV-files}

\item{write_to_file_path}{File path where to write the files, defaults to the working directory}
}
\value{
Returns a data.table in EARS-Net format
}
\description{
Note: In 2025, changes with EpiPulse cases expected to the EARS-Net metadata, thus the conversion is an interim version.
}
\examples{
From_EHRBSI_Convertor(input_file_path = "C:/DUMMY_DATA/PATH")
}
