% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_commensal_summary}
\alias{generate_commensal_summary}
\title{Generate commensal summary text}
\usage{
generate_commensal_summary(episodes_data, for_markdown = FALSE)
}
\arguments{
\item{episodes_data}{Data frame with episodes data}

\item{for_markdown}{If TRUE return markdown, otherwise HTML}
}
\value{
Character string
}
\description{
Generate commensal summary text
}
