% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_demographics_summary}
\alias{generate_demographics_summary}
\title{Generate demographics summary HTML}
\usage{
generate_demographics_summary(patient_data, for_markdown = FALSE)
}
\arguments{
\item{patient_data}{Patient data frame}
}
\value{
HTML string for Shiny UI or character for markdown
}
\description{
Generate demographics summary HTML
}
