% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_episodes_composition_summary}
\alias{generate_episodes_composition_summary}
\title{Generate episodes composition summary HTML}
\usage{
generate_episodes_composition_summary(episodes_data, for_markdown = FALSE)
}
\arguments{
\item{episodes_data}{Episodes data frame}
}
\value{
HTML string for Shiny UI or character for markdown
}
\description{
Generate episodes composition summary HTML
}
