% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_episodes_type_summary}
\alias{generate_episodes_type_summary}
\title{Generate episodes type summary HTML}
\usage{
generate_episodes_type_summary(
  episodes_data,
  patient_data,
  for_markdown = FALSE
)
}
\arguments{
\item{episodes_data}{Episodes data frame}

\item{patient_data}{Patient data frame (for patient count)}
}
\value{
HTML string for Shiny UI or character for markdown
}
\description{
Generate episodes type summary HTML
}
