% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_healthcare_facilities_summary}
\alias{generate_healthcare_facilities_summary}
\title{Generate healthcare facilities summary HTML}
\usage{
generate_healthcare_facilities_summary(
  processed_data_stats,
  raw_data_stats,
  for_markdown = FALSE
)
}
\arguments{
\item{processed_data_stats}{List with processed data statistics}

\item{raw_data_stats}{List with raw data statistics}
}
\value{
HTML string for Shiny UI or character for markdown
}
\description{
Generate healthcare facilities summary HTML
}
