% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_infection_type_summary}
\alias{generate_infection_type_summary}
\title{Generate infection type summary HTML}
\usage{
generate_infection_type_summary(episodes_data, for_markdown = FALSE)
}
\arguments{
\item{episodes_data}{Episodes data frame}
}
\value{
HTML string for Shiny UI or character for markdown
}
\description{
Generate infection type summary HTML
}
