% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_processed_data_summary}
\alias{generate_processed_data_summary}
\title{Generate processed data summary HTML}
\usage{
generate_processed_data_summary(processed_data_stats, for_markdown = FALSE)
}
\arguments{
\item{processed_data_stats}{List with processed data statistics}
}
\value{
HTML string for Shiny UI or character for markdown
}
\description{
Generate processed data summary HTML
}
