% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_raw_data_summary}
\alias{generate_raw_data_summary}
\title{Generate raw data summary HTML}
\usage{
generate_raw_data_summary(raw_data_stats, for_markdown = FALSE)
}
\arguments{
\item{raw_data_stats}{List with raw data statistics}
}
\value{
HTML string for Shiny UI or character for markdown
}
\description{
Generate raw data summary HTML
}
