% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{generate_timeseries_summary}
\alias{generate_timeseries_summary}
\title{Generate time series summary text}
\usage{
generate_timeseries_summary(
  episode_summary,
  aggregation = "month",
  for_markdown = FALSE
)
}
\arguments{
\item{episode_summary}{Data frame with EpisodeStartDate and EpisodeId}

\item{aggregation}{Time aggregation level (default "month")}

\item{for_markdown}{If TRUE return markdown, otherwise HTML}
}
\value{
Character string
}
\description{
Generate time series summary text
}
