% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{get_column_or_default}
\alias{get_column_or_default}
\title{Get column from data frame or return default value}
\usage{
get_column_or_default(data, col, default)
}
\arguments{
\item{data}{Data frame to extract from}

\item{col}{Column name to extract}

\item{default}{Default value if column doesn't exist}
}
\value{
Column values or default value
}
\description{
Get column from data frame or return default value
}
