% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{get_country_lookups}
\alias{get_country_lookups}
\title{Get lookup tables for a country}
\usage{
get_country_lookups(country_code)
}
\arguments{
\item{country_code}{Two-letter country code}
}
\value{
List of lookup vectors for the specified country
}
\description{
Get lookup tables for a country
}
