% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{get_specialty_column}
\alias{get_specialty_column}
\title{Get specialty column name from patient data}
\usage{
get_specialty_column(pat)
}
\arguments{
\item{pat}{Patient data frame}
}
\value{
String with column name or NULL if not found
}
\description{
Get specialty column name from patient data
}
