% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{has_column}
\alias{has_column}
\title{Check if a column exists in data frame}
\usage{
has_column(data, col)
}
\arguments{
\item{data}{Data frame to check}

\item{col}{Column name to check for}
}
\value{
Logical indicating if column exists
}
\description{
Check if a column exists in data frame
}
