% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ID_Create.R
\name{ID_Create}
\alias{ID_Create}
\title{Helper function for complex ID creation}
\usage{
ID_Create(DT, cols, new_col = "ParentId")
}
\arguments{
\item{DT}{Input data.table}

\item{cols}{Columns to include into the new ID}

\item{new_col}{Name of the new ID column}
}
\value{
Returns the data with the newly defined ID
}
\description{
Helper function for complex ID creation
}
