% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{initialize_resistance_columns}
\alias{initialize_resistance_columns}
\title{Initialize standard resistance table columns}
\usage{
initialize_resistance_columns(res_data, mechanism_cols = c())
}
\arguments{
\item{res_data}{Base resistance data frame}

\item{mechanism_cols}{Optional vector of mechanism result column names}
}
\value{
Data frame with all standard resistance columns initialized
}
\description{
Initialize standard resistance table columns
}
