% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{load_shared_lookups}
\alias{load_shared_lookups}
\title{Load shared lookup tables (if available)}
\usage{
load_shared_lookups()
}
\description{
Load shared lookup tables (if available)
}
\keyword{internal}
