% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\name{migrate_all_v1_to_v2}
\alias{migrate_all_v1_to_v2}
\title{Migrate EE/MT v1 configs to v2 and write side-by-side files}
\usage{
migrate_all_v1_to_v2()
}
\description{
Migrate EE/MT v1 configs to v2 and write side-by-side files
}
