% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_age_distribution}
\alias{plot_age_distribution}
\title{Plot age distribution pie chart}
\usage{
plot_age_distribution(patient_data)
}
\arguments{
\item{patient_data}{Patient data frame with Age column}
}
\value{
ggplot2 object
}
\description{
Plot age distribution pie chart
}
