% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_commensal_pie}
\alias{plot_commensal_pie}
\title{Plot commensal pie chart}
\usage{
plot_commensal_pie(episodes_data, filter_class = NULL)
}
\arguments{
\item{episodes_data}{Data frame with episodes data}

\item{filter_class}{Optional episode class filter ("HA", "CA", or NULL for all)}
}
\value{
ggplot2 object
}
\description{
Plot commensal pie chart
}
