% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_data_cleaning_pie}
\alias{plot_data_cleaning_pie}
\title{Plot data cleaning pie chart}
\usage{
plot_data_cleaning_pie(raw_data_stats, processed_data_stats)
}
\arguments{
\item{raw_data_stats}{List with raw data statistics (total_records, total_patients)}

\item{processed_data_stats}{List with processed data statistics (final_isolates, contaminants_removed)}
}
\value{
ggplot2 object
}
\description{
Plot data cleaning pie chart
}
