% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_episode_composition}
\alias{plot_episode_composition}
\title{Plot episode composition pie chart}
\usage{
plot_episode_composition(episodes_data, filter_class = NULL)
}
\arguments{
\item{episodes_data}{Data frame with episodes data}

\item{filter_class}{Optional filter for episode class ("HA", "CA", or NULL for all)}
}
\value{
ggplot2 object
}
\description{
Plot episode composition pie chart
}
