% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_gender_distribution}
\alias{plot_gender_distribution}
\title{Plot gender distribution pie chart}
\usage{
plot_gender_distribution(patient_data)
}
\arguments{
\item{patient_data}{Patient data frame with Sex column}
}
\value{
ggplot2 object
}
\description{
Plot gender distribution pie chart
}
