% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_hospital_ward_episodes_total}
\alias{plot_hospital_ward_episodes_total}
\title{Plot total episodes by ward}
\usage{
plot_hospital_ward_episodes_total(episodes_with_ward)
}
\arguments{
\item{episodes_with_ward}{Episodes data merged with ward information}
}
\value{
ggplot2 object
}
\description{
Plot total episodes by ward
}
