% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_hospital_ward_episodes_types}
\alias{plot_hospital_ward_episodes_types}
\title{Plot episode types by ward (stacked bar chart)}
\usage{
plot_hospital_ward_episodes_types(episodes_with_ward)
}
\arguments{
\item{episodes_with_ward}{Episodes data merged with ward information}
}
\value{
ggplot2 object
}
\description{
Plot episode types by ward (stacked bar chart)
}
