% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_infection_type}
\alias{plot_infection_type}
\title{Plot infection type pie chart}
\usage{
plot_infection_type(episodes_data, filter_class = NULL)
}
\arguments{
\item{episodes_data}{Data frame with episodes data}

\item{filter_class}{Optional filter for episode class ("HA", "CA", or NULL for all)}
}
\value{
ggplot2 object
}
\description{
Plot infection type pie chart
}
