% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_monomicrobial_pathogens}
\alias{plot_monomicrobial_pathogens}
\title{Plot top 20 pathogens in monomicrobial episodes}
\usage{
plot_monomicrobial_pathogens(episode_summary)
}
\arguments{
\item{episode_summary}{Data frame with episode-level summary (one row per episode)}
}
\value{
ggplot2 object
}
\description{
Plot top 20 pathogens in monomicrobial episodes
}
