% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_polymicrobial_combinations}
\alias{plot_polymicrobial_combinations}
\title{Plot pathogen combinations in polymicrobial episodes}
\usage{
plot_polymicrobial_combinations(episode_summary)
}
\arguments{
\item{episode_summary}{Data frame with episode-level summary}
}
\value{
ggplot2 object
}
\description{
Plot pathogen combinations in polymicrobial episodes
}
