% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_polymicrobial_individual}
\alias{plot_polymicrobial_individual}
\title{Plot individual pathogens in polymicrobial episodes}
\usage{
plot_polymicrobial_individual(episode_summary)
}
\arguments{
\item{episode_summary}{Data frame with episode-level summary}
}
\value{
ggplot2 object
}
\description{
Plot individual pathogens in polymicrobial episodes
}
