% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_seasonality}
\alias{plot_seasonality}
\title{Plot seasonal pattern (average episodes by month of year)}
\usage{
plot_seasonality(episode_summary)
}
\arguments{
\item{episode_summary}{Data frame with EpisodeStartDate and EpisodeId}
}
\value{
ggplot2 object
}
\description{
Plot seasonal pattern (average episodes by month of year)
}
