% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_specialty_per_episode}
\alias{plot_specialty_per_episode}
\title{Plot specialty distribution pie chart (per episode)}
\usage{
plot_specialty_per_episode(episodes_data, patient_data, specialty_col = NULL)
}
\arguments{
\item{episodes_data}{Episodes data frame}

\item{patient_data}{Patient data frame}

\item{specialty_col}{Specialty column name}
}
\value{
ggplot2 object
}
\description{
Plot specialty distribution pie chart (per episode)
}
