% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_specialty_per_patient}
\alias{plot_specialty_per_patient}
\title{Plot specialty distribution pie chart (per patient)}
\usage{
plot_specialty_per_patient(patient_data, specialty_col = NULL)
}
\arguments{
\item{patient_data}{Patient data frame}

\item{specialty_col}{Specialty column name}
}
\value{
ggplot2 object
}
\description{
Plot specialty distribution pie chart (per patient)
}
