% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportModules.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot episodes over time}
\usage{
plot_timeseries(episode_summary, aggregation = "month", show_trend = TRUE)
}
\arguments{
\item{episode_summary}{Data frame with EpisodeStartDate and EpisodeId}

\item{aggregation}{Time aggregation level (default "month")}

\item{show_trend}{Whether to overlay a LOESS trend line}
}
\value{
ggplot2 object
}
\description{
Plot episodes over time
}
