% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibioticHAI_longShort.r
\name{recode_to_HAI_short}
\alias{recode_to_HAI_short}
\title{Recode HAI long names to HAI short codes}
\usage{
recode_to_HAI_short(
  data,
  metadata_path,
  long_col = "Antibiotic",
  short_col = "Antibiotic",
  sheet = "Coded values",
  coded_list_id = "AntibioticHAI"
)
}
\arguments{
\item{data}{Data frame containing HAI long antibiotic names}

\item{metadata_path}{Path to the metadata Excel file containing coded values}

\item{long_col}{Column name containing HAI long names}

\item{short_col}{Column name to create for HAI short codes}

\item{sheet}{Sheet name in the Excel file}

\item{coded_list_id}{Identifier for the antibiotic coded list}
}
\value{
Data frame with HAI short codes added
}
\description{
Recode HAI long names to HAI short codes
}
\keyword{internal}
