% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shorten_specialty_names}
\alias{shorten_specialty_names}
\title{Shorten specialty names for plot labels}
\usage{
shorten_specialty_names(specialties)
}
\arguments{
\item{specialties}{Character vector of specialty names}
}
\value{
Character vector with shortened names (newlines for wrapping)
}
\description{
Shorten specialty names for plot labels
}
