% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{standardize_all_table_dates}
\alias{standardize_all_table_dates}
\title{Standardize date formats across all EHR-BSI tables}
\usage{
standardize_all_table_dates(result_list)
}
\arguments{
\item{result_list}{List containing ehrbsi, patient, isolate, res tables}
}
\value{
List with all date columns in all tables converted to 'yyyy-mm-dd' format
}
\description{
Applies standardize_date_format to all tables in the result list.
}
