% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{standardize_all_table_mic_sign}
\alias{standardize_all_table_mic_sign}
\title{Standardize MICSusceptibilitySign values across all EHR-BSI tables}
\usage{
standardize_all_table_mic_sign(result_list)
}
\arguments{
\item{result_list}{List containing ehrbsi, patient, isolate, res tables}
}
\value{
List with MICSusceptibilitySign '<' values changed to '<='
}
\description{
Applies standardize_mic_susceptibility_sign to all tables in the result list.
}
