% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{standardize_all_table_sex}
\alias{standardize_all_table_sex}
\title{Standardize Sex values across all EHR-BSI tables}
\usage{
standardize_all_table_sex(result_list)
}
\arguments{
\item{result_list}{List containing ehrbsi, patient, isolate, res tables}
}
\value{
List with Sex values standardized to 'F', 'M', or 'OTH'
}
\description{
Applies standardize_sex_values to all tables in the result list.
}
