% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodingFuncs.R
\name{standardize_all_table_unit_specialty}
\alias{standardize_all_table_unit_specialty}
\title{Standardize UnitSpecialtyShort values across all EHR-BSI tables}
\usage{
standardize_all_table_unit_specialty(result_list)
}
\arguments{
\item{result_list}{List containing ehrbsi, patient, isolate, res tables}
}
\value{
List with UnitSpecialtyShort '0' and 'O' values changed to 'OTH'
}
\description{
Applies standardize_unit_specialty_short to all tables in the result list.
}
