% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_date}
\alias{to_date}
\title{Robust date coercion}
\usage{
to_date(x)
}
\arguments{
\item{x}{Value(s) to coerce to Date}
}
\value{
Date vector
}
\description{
Supports Date/POSIX, ISO, EU/US formats, with/without time, and Excel serials.
}
