% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UNK_O_Handler.R
\name{UNK_O_Handler}
\alias{UNK_O_Handler}
\title{Helper function for "UNK"/"N/A"/"O" recodes}
\usage{
UNK_O_Handler(DT, cols = NULL)
}
\arguments{
\item{DT}{Input data.table}

\item{cols}{Columns to recode; defaults to all character columns}
}
\value{
Returns the data with relevant recodes
}
\description{
Helper function for "UNK"/"N/A"/"O" recodes
}
