% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseDashboard.r
\name{visual_bsi_dashboard}
\alias{visual_bsi_dashboard}
\title{Shiny-based Interactive BSI Dashboard}
\usage{
visual_bsi_dashboard(data)
}
\arguments{
\item{data}{A list of data.frames. Expected to have an \code{ehrbsi} data.frame.}
}
\value{
Launches a Shiny application.
}
\description{
This function creates a Shiny app with full interactive controls for
exploring BSI data with dropdowns, sliders, and filters.
}
