% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualFuncs.r
\name{visual_ho_bsi_by_hospital}
\alias{visual_ho_bsi_by_hospital}
\title{Hospital-Onset BSI Distribution by Hospital}
\usage{
visual_ho_bsi_by_hospital(data)
}
\arguments{
\item{data}{A list of data.frames. Expected to have an \code{ehrbsi} data.frame
with \code{HospitalId} and \code{NumberOfHOHABSIs} columns.}
}
\value{
A ggplot object for rendering.
}
\description{
This visual function creates a bar chart showing the number of Hospital-Onset
Healthcare-Associated BSIs (HO-HA BSIs) per hospital.
}
