% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryConfigs.R
\docType{data}
\name{COUNTRY_CONFIGS}
\alias{COUNTRY_CONFIGS}
\title{Country Configuration Objects}
\format{
List of configuration objects, one per country
}
\usage{
COUNTRY_CONFIGS
}
\description{
Country Configuration Objects
}
\keyword{datasets}
